+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Changed
 
+- Moved project data for Groundwork from extras jsonb field and a text field to normalized tables [#5286](https://github.com/raster-foundry/raster-foundry/pull/5286)
 - Clear Python Lambda function pip cache [#5274](https://github.com/raster-foundry/raster-foundry/pull/5274)
 
 ### Deprecated